+++ gdbgui/src/js/ControlButtons.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import Actions from "./Actions.js";
-import GdbApi from "./GdbApi.jsx";
+import GdbApi from "./GdbApi";
 import { store } from "statorgfc";
 
 class ControlButtons extends React.Component {
@@ -40,7 +40,7 @@ class ControlButtons extends React.Component {
         <button
           onClick={() => Actions.send_signal("SIGINT", this.state.gdb_pid)}
           type="button"
-          title="Send Interrupt signal (SIGINT) to gdb process to pause it and allow interaction with it"
+          title="Send Interrupt signal (SIGINT) to gdb process to pause it (if it's running)"
           className={btn_class}
         >
           <span className="glyphicon glyphicon-pause" />