+++ gdbgui/src/js/GdbApi.tsx
@@ -10,12 +10,18 @@ import GdbVariable from "./GdbVariable.jsx";
 import constants from "./constants.js";
 import process_gdb_response from "./process_gdb_response.js";
 import React from "react";
+import io from "socket.io-client";
 void React; // needed when using JSX, but not marked as used
-
 /* global debug */
 
 // print to console if debug is true
-let debug_print;
+let debug_print: {
+  (arg0: string): void;
+  (...data: any[]): void;
+  (message?: any, ...optionalParams: any[]): void;
+  (): void;
+};
+// @ts-ignore
 if (debug) {
   debug_print = console.info;
 } else {
@@ -28,81 +34,115 @@ if (debug) {
  * This object contains methods to interact with
  * gdb, but does not directly render anything in the DOM.
  */
+// @ts-ignore
+const initial_data = window.initial_data;
+let socket: SocketIOClient.Socket;
 const GdbApi = {
+  getSocket: function() {
+    return socket;
+  },
   init: function() {
     const TIMEOUT_MIN = 5;
-    /* global io */
-    /* global initial_data */
-    GdbApi.socket = io.connect(`/gdb_listener`, {
+    socket = io.connect(`/gdb_listener`, {
       timeout: TIMEOUT_MIN * 60 * 1000,
-      query: `csrf_token=${initial_data.csrf_token}&gdbpid=${initial_data.gdbpid}`
+      query: {
+        csrf_token: initial_data.csrf_token,
+        gdbpid: initial_data.gdbpid,
+        gdb_command: initial_data.gdb_command
+      }
     });
 
-    GdbApi.socket.on("connect", function() {
+    socket.on("connect", function() {
       debug_print("connected");
     });
 
-    GdbApi.socket.on("gdb_response", function(response_array) {
+    socket.on("gdb_response", function(response_array: any) {
+      // @ts-ignore
       clearTimeout(GdbApi._waiting_for_response_timeout);
       store.set("waiting_for_response", false);
       process_gdb_response(response_array);
     });
-
-    GdbApi.socket.on("error_running_gdb_command", function(data) {
+    socket.on("fatal_server_error", function(data: { message: null | string }) {
+      Actions.add_console_entries(
+        `Message from server: ${data.message}`,
+        constants.console_entry_type.STD_ERR
+      );
+      socket.close();
+    });
+    socket.on("error_running_gdb_command", function(data: { message: any }) {
       Actions.add_console_entries(
         `Error occurred on server when running gdb command: ${data.message}`,
         constants.console_entry_type.STD_ERR
       );
+      socket.close();
     });
 
-    GdbApi.socket.on("server_error", function(data) {
+    socket.on("server_error", function(data: { message: any }) {
       Actions.add_console_entries(
         `Server message: ${data.message}`,
         constants.console_entry_type.STD_ERR
       );
     });
 
-    GdbApi.socket.on("gdb_pid", function(gdb_pid_obj) {
-      let gdb_pid = gdb_pid_obj.pid,
-        message = gdb_pid_obj.message,
-        error = gdb_pid_obj.error,
-        using_existing = gdb_pid_obj.using_existing;
+    socket.on("debug_session_connection_event", function(gdb_pid_obj: {
+      pid: number;
+      message: string | void;
+      ok: boolean;
+      started_new_gdb_process: boolean;
+    }) {
+      const gdb_pid = gdb_pid_obj.pid;
+      const message = gdb_pid_obj.message;
+      const error = !gdb_pid_obj.ok;
+      const started_new_gdb_process = gdb_pid_obj.started_new_gdb_process;
 
-      Actions.add_console_entries(
-        message,
-        error
-          ? constants.console_entry_type.STD_ERR
-          : constants.console_entry_type.GDBGUI_OUTPUT
-      );
-
+      if (message) {
+        Actions.add_console_entries(
+          message,
+          error
+            ? constants.console_entry_type.STD_ERR
+            : constants.console_entry_type.GDBGUI_OUTPUT
+        );
+      }
+      if (error) {
+        socket.close();
+        return;
+      }
       store.set("gdb_pid", gdb_pid);
-      Actions.add_console_entries(
-        `${store.get("interpreter")} process ${gdb_pid} is running for this tab`,
-        constants.console_entry_type.GDBGUI_OUTPUT
-      );
-      if (using_existing) {
-        Actions.refresh_state_for_gdb_pause();
-      } else {
+
+      if (started_new_gdb_process) {
         GdbApi.run_initial_commands();
+      } else {
+        Actions.refresh_state_for_gdb_pause();
       }
     });
 
-    GdbApi.socket.on("disconnect", function() {
+    socket.on("disconnect", function() {
       // we no longer need to warn the user before they exit the page since the gdb process
       // on the server is already gone
       window.onbeforeunload = () => null;
 
-      // show modal
       Actions.show_modal(
         "",
-        <span>
-          The gdbgui server has shutdown. This tab will no longer function as expected.
-        </span>
+        <>
+          <p>
+            The connection to the gdb session has been closed. This tab will no longer
+            function as expected.
+          </p>
+          <p className="font-bold">
+            To start a new session or connect to a different session, go to the{" "}
+            <a href="/dashboard">dashboard</a>.
+          </p>
+        </>
       );
-      debug_print("disconnected");
-      if (debug) {
-        window.location.reload(true);
-      }
+      Actions.add_console_entries(
+        `The connection to the gdb session has been closed. To start a new session, go to ${window.location.origin}/dashboard`,
+        constants.console_entry_type.STD_ERR
+      );
+
+      // @ts-ignore
+      // if (debug) {
+      //   window.location.reload(true);
+      // }
     });
   },
   _waiting_for_response_timeout: null,
@@ -111,13 +151,10 @@ const GdbApi = {
     GdbApi.run_gdb_command("-exec-run");
   },
   run_initial_commands: function() {
-    const cmds = ["set breakpoint pending on"];
+    const cmds = ["-list-features", "-list-target-features"];
     for (const src in initial_data.remap_sources) {
       const dst = initial_data.remap_sources[src];
       cmds.push(`set substitute-path "${src}" "${dst}"`);
-    }
-    if (initial_data.initial_gdb_user_command) {
-      cmds.push(initial_data.initial_gdb_user_command);
     }
     GdbApi.run_gdb_command(cmds);
   },
@@ -172,11 +209,13 @@ const GdbApi = {
     Actions.inferior_program_resuming();
     GdbApi.run_gdb_command("-exec-interrupt");
   },
-  send_autocomplete_command: function(command) {
+  send_autocomplete_command: function(command: string) {
     Actions.inferior_program_resuming();
     GdbApi.run_gdb_command("complete " + command);
   },
-  click_gdb_cmd_button: function(e) {
+  click_gdb_cmd_button: function(e: {
+    currentTarget: { dataset: { [x: string]: any; cmd: undefined; cmd0: undefined } };
+  }) {
     if (e.currentTarget.dataset.cmd !== undefined) {
       // run single command
       // i.e. <a data-cmd='cmd' />
@@ -201,12 +240,12 @@ const GdbApi = {
       );
     }
   },
-  select_frame: function(framenum) {
+  select_frame: function(framenum: any) {
     // TODO this command is deprecated (https://sourceware.org/gdb/onlinedocs/gdb/GDB_002fMI-Stack-Manipulation.html)
     // This command in deprecated in favor of passing the ‘--frame’ option to every command.
     GdbApi.run_command_and_refresh_state(`-stack-select-frame ${framenum}`);
   },
-  select_thread_id: function(thread_id) {
+  select_thread_id: function(thread_id: any) {
     // TODO this command is deprecated (http://www.sourceware.org/gdb/current/onlinedocs/gdb/GDB_002fMI-Thread-Commands.html)
     // This command is deprecated in favor of explicitly using the ‘--thread’ option to each command.
     GdbApi.run_command_and_refresh_state(`-thread-select ${thread_id}`);
@@ -218,10 +257,15 @@ const GdbApi = {
   waiting_for_response: function() {
     store.set("waiting_for_response", true);
     const WAIT_TIME_SEC = 10;
+    // @ts-ignore
     clearTimeout(GdbApi._waiting_for_response_timeout);
+    // @ts-ignore
     GdbApi._waiting_for_response_timeout = setTimeout(() => {
       Actions.clear_program_state();
       store.set("waiting_for_response", false);
+      if (GdbApi.getSocket().disconnected) {
+        return;
+      }
 
       Actions.add_console_entries(
         `No gdb response received after ${WAIT_TIME_SEC} seconds.`,
@@ -255,37 +299,41 @@ const GdbApi = {
    * @param cmd: a string or array of strings, that are directly evaluated by gdb
    * @return nothing
    */
-  run_gdb_command: function(cmd) {
+  run_gdb_command: function(cmd: any) {
+    // @ts-ignore
     if (_.trim(cmd) === "") {
       return;
     }
 
     let cmds = cmd;
+    // @ts-ignore
     if (_.isString(cmds)) {
       cmds = [cmds];
     }
-
     // add the send command to the console to show commands that are
     // automatically run by gdb
     if (store.get("show_all_sent_commands_in_console")) {
       Actions.add_console_entries(cmds, constants.console_entry_type.SENT_COMMAND);
     }
 
-    GdbApi.waiting_for_response();
-    GdbApi.socket.emit("run_gdb_command", { cmd: cmds });
+    if (socket.connected) {
+      GdbApi.waiting_for_response();
+      socket.emit("run_gdb_command", { cmd: cmds });
+    }
   },
   /**
    * Run a user-defined command, then refresh the store
    * @param user_cmd (str or array): command or commands to run before refreshing store
    */
-  run_command_and_refresh_state: function(user_cmd) {
+  run_command_and_refresh_state: function(user_cmd: string | any[]) {
     // if(!user_cmd){
     //     console.error('missing required argument')
     //     return
     // }
-    let cmds = [];
-    if (_.isArray(user_cmd)) {
+    let cmds: any[] = [];
+    if (Array.isArray(user_cmd)) {
       cmds = cmds.concat(user_cmd);
+      // @ts-ignore
     } else if (_.isString(user_cmd) && user_cmd.length > 0) {
       cmds.push(user_cmd);
     }
@@ -312,14 +360,8 @@ const GdbApi = {
       // and the name and type for arrays, structures and unions.
       constants.IGNORE_ERRORS_TOKEN_STR + "-stack-list-variables --simple-values"
     ];
-    if (store.get("interpreter") === "gdb") {
-      // update all user-defined variables in gdb
-      cmds.push(constants.IGNORE_ERRORS_TOKEN_STR + "-var-update --all-values *");
-    } else if (store.get("interpreter") === "lldb") {
-      // the * arg doesn't work, so loop over all
-      // names and push commands for each
-      cmds = cmds.concat(GdbVariable.get_update_cmds());
-    }
+    // update all user-defined variables in gdb
+    cmds.push(constants.IGNORE_ERRORS_TOKEN_STR + "-var-update --all-values *");
 
     // update registers
     cmds = cmds.concat(Registers.get_update_cmds());
@@ -338,9 +380,9 @@ const GdbApi = {
   refresh_breakpoints: function() {
     GdbApi.run_gdb_command([GdbApi.get_break_list_cmd()]);
   },
-  get_inferior_binary_last_modified_unix_sec(path) {
+  get_inferior_binary_last_modified_unix_sec(path: any) {
     $.ajax({
-      beforeSend: function(xhr) {
+      beforeSend: function(xhr: { setRequestHeader: (arg0: string, arg1: any) => void }) {
         xhr.setRequestHeader("x-csrftoken", initial_data.csrf_token);
       },
       url: "/get_last_modified_unix_sec",
@@ -351,31 +393,16 @@ const GdbApi = {
       error: GdbApi._error_getting_last_modified_unix_sec
     });
   },
-  get_insert_break_cmd: function(fullname, line) {
-    if (store.get("interpreter") === "gdb") {
-      return [`-break-insert "${fullname}:${line}"`];
-    } else {
-      console.log("TODOLLDB - find mi-friendly command");
-      return [`breakpoint set --file ${fullname} --line ${line}`];
-    }
+  get_insert_break_cmd: function(fullname: any, line: any) {
+    return [`-break-insert "${fullname}:${line}"`];
   },
-  get_delete_break_cmd: function(bkpt_num) {
-    if (store.get("interpreter") === "gdb") {
-      return `-break-delete ${bkpt_num}`;
-    } else {
-      console.log("TODOLLDB - find mi-friendly command");
-      return `breakpoint delete ${bkpt_num}`;
-    }
+  get_delete_break_cmd: function(bkpt_num: any) {
+    return `-break-delete ${bkpt_num}`;
   },
   get_break_list_cmd: function() {
-    if (store.get("interpreter") === "gdb") {
-      return "-break-list";
-    } else if (store.get("interpreter") === "lldb") {
-      console.log("TODOLLDB - find mi-friendly command");
-      return "breakpoint list";
-    }
+    return "-break-list";
   },
-  get_load_binary_and_arguments_cmds(binary, args) {
+  get_load_binary_and_arguments_cmds(binary: any, args: any) {
     // tell gdb which arguments to use when calling the binary, before loading the binary
     let cmds = [
       `-exec-arguments ${args}`, // Set the inferior program arguments, to be used in the next `-exec-run`
@@ -388,10 +415,10 @@ const GdbApi = {
     cmds.push(GdbApi.get_break_list_cmd());
     return cmds;
   },
-  set_assembly_flavor(flavor) {
-    GdbApi.run_gdb_command("set disassembly-flavor " + flavor);
+  set_assembly_flavor(flavor: string) {
+    GdbApi.run_gdb_command(`set disassembly-flavor ${flavor}`);
   },
-  _recieve_last_modified_unix_sec(data) {
+  _recieve_last_modified_unix_sec(data: { path: any; last_modified_unix_sec: any }) {
     if (data.path === store.get("inferior_binary_path")) {
       store.set(
         "inferior_binary_path_last_modified_unix_sec",
@@ -399,10 +426,11 @@ const GdbApi = {
       );
     }
   },
-  _error_getting_last_modified_unix_sec(data) {
+  _error_getting_last_modified_unix_sec(data: any) {
     void data;
     store.set("inferior_binary_path", null);
   }
 };
-
+// @ts-ignore
+GdbApi.socket = socket;
 export default GdbApi;