+++ noxfile.py
@@ -1,7 +1,9 @@
 import nox  # type: ignore
 from pathlib import Path
 from sys import platform
+import subprocess
 
+nox.options.reuse_existing_virtualenvs
 nox.options.sessions = ["tests", "lint", "docs"]
 python = ["3.6", "3.7", "3.8"]
 
@@ -9,21 +11,30 @@ python = ["3.6", "3.7", "3.8"]
 doc_dependencies = [".", "mkdocs", "mkdocs-material"]
 lint_dependencies = ["black", "flake8", "mypy", "check-manifest"]
 files_to_lint = ["gdbgui", "tests"] + [str(p) for p in Path(".").glob("*.py")]
+publish_deps = ["setuptools", "wheel", "twine"]
 
 
-@nox.session(reuse_venv=True, python=python)
-def tests(session):
+@nox.session(reuse_venv=True)
+def python_tests(session):
     session.install(".", "pytest", "pytest-cov")
     tests = session.posargs or ["tests"]
     session.run(
         "pytest", "--cov=gdbgui", "--cov-config", ".coveragerc", "--cov-report=", *tests
     )
+    session.notify("cover")
 
+
+@nox.session(reuse_venv=True)
+def js_tests(session):
     session.run("yarn", "install", external=True)
     session.run("yarn", "test", external=True)
     session.run("yarn", "build", external=True)
 
-    session.notify("cover")
+
+@nox.session(reuse_venv=True, python=python)
+def tests(session):
+    python_tests(session)
+    js_tests(session)
 
 
 @nox.session(reuse_venv=True)
@@ -35,13 +46,22 @@ def cover(session):
         "report",
         "--show-missing",
         "--omit=gdbgui/SSLify.py",
-        "--fail-under=30",
+        "--fail-under=20",
     )
     session.run("coverage", "erase")
 
 
 @nox.session(reuse_venv=True)
 def lint(session):
+
+    session.install(".", *lint_dependencies)
+    session.run("black", "--check", *files_to_lint)
+    session.run("flake8", *files_to_lint)
+    session.run("mypy", *files_to_lint)
+    session.run(
+        "check-manifest", "--ignore", "gdbgui/static/js/*",
+    )
+    session.run("python", "setup.py", "check", "--metadata", "--strict")
     session.run(
         "npx",
         "prettier@1.18.2",
@@ -51,16 +71,6 @@ def lint(session):
         "gdbgui/src/js/**/*",
         external=True,
     )
-    session.install(*lint_dependencies)
-    session.run("black", "--check", *files_to_lint)
-    session.run("flake8", *files_to_lint)
-    session.run("mypy", *files_to_lint)
-    session.run(
-        "check-manifest",
-        "--ignore",
-        "build.js,gdbgui/static/js,gdbgui/static/js/build.js.map",
-    )
-    session.run("python", "setup.py", "check", "--metadata", "--strict")
 
 
 @nox.session(reuse_venv=True)
@@ -84,19 +94,25 @@ def docs(session):
     session.run("mkdocs", "build")
 
 
-@nox.session(reuse_venv=True, python=python)
+@nox.session(reuse_venv=True)
 def develop(session):
-    session.install(*doc_dependencies, *lint_dependencies)
     session.install("-e", ".")
-    command = "source %s/bin/activate" % (session.virtualenv.location_name)
-    session.log("Virtual Environment is ready to be used for development")
-    session.log("To use, run: '%s'", command)
+    session.run("yarn", "install", external=True)
+    print("Watching JavaScript file and Python files for changes")
+    with subprocess.Popen(["yarn", "start"]):
+        session.run("python", "gdbgui/backend.py")
+
+
+@nox.session(reuse_venv=True)
+def serve(session):
+    session.install("-e", ".")
+    session.run("python", "gdbgui/backend.py", *session.posargs)
 
 
 @nox.session(reuse_venv=True)
 def build(session):
-    session.install("setuptools", "wheel", "twine")
-    session.run("rm", "-rf", "dist", external=True)
+    session.install(*publish_deps)
+    session.run("rm", "-rf", "dist", "build", external=True)
     session.run("yarn", "build", external=True)
     session.run("python", "setup.py", "--quiet", "sdist", "bdist_wheel")
     session.run("twine", "check", "dist/*")
@@ -104,6 +120,7 @@ def build(session):
 
 @nox.session(reuse_venv=True)
 def publish(session):
+    session.install(*publish_deps)
     build(session)
     print("REMINDER: Has the changelog been updated?")
     session.run("python", "-m", "twine", "upload", "dist/*")